download test features
